🐛:: change type

+++ data/src/main/java/com/semicolon/data/remote/response/challenge/ChallengeListResponse.kt
@@ -11,7 +11,7 @@ data class ChallengeListResponse(
     @SerializedName("challenge_list") val challengeList: List<ChallengeResponse>
 ) {
     data class ChallengeResponse(
-        @SerializedName("id") val id: Long,
+        @SerializedName("id") val id: Int,
         @SerializedName("name") val name: String,
         @SerializedName("start_at") val startAt: String,
         @SerializedName("end_at") val endAt: String,
@@ -26,7 +26,7 @@ data class ChallengeListResponse(
     )
 
     data class User(
-        @SerializedName("user_id") val userId: Long,
+        @SerializedName("user_id") val userId: Int,
         @SerializedName("name") val name: String,
         @SerializedName("profile_image_url") val profileImageUrl: String
     )
